moved common graphics rendering to "DefaultEntity". Npc and Enemy are still missing
