+++ webapp/src/app/services/phaser/entities/registry/destructible.ts
@@ -1,57 +1,25 @@
-import { Point, Point3 } from '../../../../models/cross-code-map';
+import { Point3 } from '../../../../models/cross-code-map';
 import { Globals } from '../../../globals';
 import { Helper } from '../../helper';
-import { CCEntity, EntityAttributes, Fix, ScaleSettings } from '../cc-entity';
-import { Anims, AnimSheet } from '../../sheet-parser';
+import { ScaleSettings } from '../cc-entity';
+import { Anims } from '../../sheet-parser';
+import { DefaultEntity } from './default-entity';
 
-export class Destructible extends CCEntity {
-	private attributes: EntityAttributes = {
-		desType: {
-			type: 'String',
-			description: 'Type of destructible object',
-			withNull: true,
-		},
-		enemyInfo: {
-			type: 'EnemyType',
-			description: 'Enemy to spawn after destruction',
-			withNull: true,
-		},
-		spawnCondition: {
-			type: 'VarCondition',
-			description: 'Condition for Enemy to spawn',
-		},
-		blockNavMap: {
-			type: 'Boolean',
-			description: 'If true, block path map and update when destroyed',
-		},
-		permaDestruct: {
-			type: 'Boolean',
-			description: 'If true, then destructible stays destroyed after reentering map',
-		},
-		onDestructIncrease: {
-			type: 'VarName',
-			description: 'Variable to increase by one when destroyed',
-		},
-		onPreDestructIncrease: {
-			type: 'VarName',
-			description: 'Variable to increase by one when destroyed'
-		}
-	};
+export class Destructible extends DefaultEntity {
 	
-	public getAttributes(): EntityAttributes {
-		return this.attributes;
-	}
-	
-	public getScaleSettings(): ScaleSettings | undefined {
+	public override getScaleSettings(): ScaleSettings | undefined {
 		return undefined;
 	}
 	
-	protected async setupType(settings: any): Promise<void> {
+	protected override async setupType(settings: any): Promise<void> {
 		const types = await Globals.jsonLoader.loadJsonMerged<DestructibleTypes>('destructible-types.json');
 		
-		this.attributes['desType'].options = {};
-		for (const name of Object.keys(types)) {
-			this.attributes['desType'].options[name] = name;
+		const attributes = this.getAttributes();
+		if (attributes['desType']) {
+			attributes['desType'].options = {};
+			for (const name of Object.keys(types)) {
+				attributes['desType'].options[name] = name;
+			}
 		}
 		
 		const type = types[settings.desType];
@@ -60,63 +28,22 @@ export class Destructible extends CCEntity {
 			return;
 		}
 		
-		const sheets: Fix[] = [];
-		const defaultSheet = type.anims.sheet as AnimSheet;
-		
-		for (const sub of (type.anims.SUB as Anims[])) {
-			const sheet = defaultSheet || type.anims.namedSheets![sub.sheet as string];
-			let frame = sub.frames ? sub.frames[0] : 0;
-			
-			// frame 1 is always glow, don't show in map editor
-			if (frame === 1) {
-				frame = 0;
-			}
-			
-			const offset: Partial<Point> = {
-				x: 0,
-				y: 0
-			};
-			if (type.anims.offset) {
-				offset.x = type.anims.offset.x;
-				offset.y = type.anims.offset.y;
-			}
-			const xCount = sheet.xCount || 999;
-			const xOffset = (frame % xCount) * sheet.width;
-			const yOffset = Math.floor(frame / xCount) * sheet.height;
-			sheets.push({
-				gfx: sheet.src,
-				x: (sheet.offX || 0) + xOffset,
-				y: (sheet.offY || 0) + yOffset,
-				w: sheet.width,
-				h: sheet.height,
-				offsetX: offset.x,
-				offsetY: offset.y,
-				renderMode: sub.renderMode
-			});
+		const anims = Helper.copy(type.anims);
+
+		// frame 1 is always glow, don't show in map editor
+		if (Array.isArray(anims.SUB)) {
+			anims.SUB = anims.SUB.filter(sub => sub.frames?.[0] !== 1);
 		}
-		
-		this.entitySettings = {
-			sheets: {
-				fix: sheets,
-			},
-			baseSize: type.size,
-		} as any;
-		
-		const mapStyle = Helper.getMapStyle(Globals.map, 'destruct');
-		for (const sheet of sheets) {
-			if (!sheet.gfx) {
-				sheet.gfx = mapStyle?.sheet;
-			}
-			const exists = await Helper.loadTexture(sheet.gfx, this.scene);
-			if (!exists) {
-				console.error('sheet does not exist: ' + sheet.gfx);
-			}
-		}
-		
-		if (sheets.length === 0) {
+
+		// sheet src is null in json; CrossCode fills it from the current map's destruct style.
+		const fallbackGfx = Helper.getMapStyle(Globals.map, 'destruct')?.sheet;
+		const ok = await this.applyAnims(anims, undefined, settings.desType, fallbackGfx);
+		if (!ok) {
 			this.generateErrorImage();
+			return;
 		}
 		
+		this.entitySettings.baseSize = type.size;
 		this.updateSettings();
 	}
